Refactor to compat
