+++ src/openscm_runner/adapters/magicc7/_compat.py
@@ -1,0 +1,13 @@
+"""
+Handling of compatibility of MAGICC imports with different states of installation
+"""
+# pylint:disable=unused-import
+try:
+    import f90nml
+    import pymagicc
+
+    HAS_PYMAGICC = True
+except ImportError:
+    f90nml = None
+    pymagicc = None
+    HAS_PYMAGICC = False
